+++ e2e/e2e_test.go
@@ -67,5 +67,5 @@ func TestMiscErrors(t *testing.T) {
 }
 
 func TestGFlowsLibs(t *testing.T) {
-	runTests(t, "./tests/gflowslibs/jsonnet/test-remote-gflowslib.yml", false)
+	runTests(t, "./tests/gflowslibs/jsonnet/*.yml", false)
 }
